Decoupled html  function calls from Vinyl class
Decoupled html  function calls from Vinyl class.
