+++ create.html
@@ -2,10 +2,11 @@
     <head>
         <script type="module">
             import {Vinyl} from "./js/vinyl.js";
+            import {renderVinylForm} from "./js/vinyl-frontend.js";
             import {ERROR} from "./js/codes.js";
 
             let vinyl = new Vinyl();
-            vinyl.renderForm("new_vinyl");
+            renderVinylForm(vinyl, "new_vinyl");
         </script>
     </head>
